scrapping/requirements.txt
asyncio
requests-html
psycopg2-binary